requirements.txt
ultralytics==8.3.175
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
PyYAML==6.0.1
